# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "description": "Social media application backend",
    "title": "PostAway"
  },
  "servers": [{ "url": "http://localhost:3200" }],
  "components": {
    "securitySchemes": {
      "JWT": { "in": "cookie", "name": "Authorization", "type": "apiKey" }
    }
  },
  "paths": {
...
    "/api/user/{id}": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get a user by Id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the user to retrieve",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User created successfully"
          },
          "500": {
            "description": "Something went wrong"
          }
        }
      }
    },
...
    "/api/user/{postId}": {
      "get": {
        "tags": ["Posts"],
        "summary": "Get posts by post id",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "ID of the post to retrieve",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post"
          },
          "500": {
            "description": "Something went wrong"
          }
        }
      }
    },
...
  }
}